IGAPP-766: rename useLocation to useUserLocation, show locationButton on timeout, navigate to bbox of city instead of current loaction
